valorant_highlight_agent/requirements.txt
opencv-python
numpy
pytube
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pyyaml
